requirements.txt
gnews==0.2.7
langchain==0.0.244
newspaper3k==0.2.8
nltk==3.7
pandas==1.4.4
pydantic==1.10.1
requests
altair<5
scikit_learn==1.3.0
streamlit==1.13.0
sumy==0.10.0


